fix Optional<Long> equality test

+++ apps/bfd-server/bfd-server-war/src/test/java/gov/cms/bfd/server/war/r4/providers/BeneficiaryTransformerV2Test.java
@@ -93,7 +93,8 @@ public final class BeneficiaryTransformerV2Test {
             .filter(
                 r ->
                     (r.getBeneficiaryId().isPresent()
-                        && r.getBeneficiaryId().get() == beneficiary.getBeneficiaryId()))
+                        && r.getBeneficiaryId().get().longValue()
+                            == beneficiary.getBeneficiaryId()))
             .collect(Collectors.toSet());
     beneficiary.getMedicareBeneficiaryIdHistories().addAll(beneficiaryMbis);
     assertThat(beneficiary, is(notNullValue()));
