+++ apps/bfd-server/bfd-server-war/src/test/java/gov/cms/bfd/server/war/stu3/providers/BeneficiaryTransformerTest.java
@@ -310,7 +310,8 @@ public final class BeneficiaryTransformerTest {
             .filter(
                 r ->
                     (r.getBeneficiaryId().isPresent()
-                        && beneficiary.getBeneficiaryId() == r.getBeneficiaryId().get()))
+                        && beneficiary.getBeneficiaryId()
+                            == r.getBeneficiaryId().get().longValue()))
             .collect(Collectors.toSet());
     beneficiary.getMedicareBeneficiaryIdHistories().addAll(beneficiaryMbis);
